Перенос cloudinary.js в tools.js в связи с реструктуризацией
